# swagger/src/main/resources/swagger-cost-seiko.yaml
openapi: 3.0.3
info:
  title: Cost seiko
  description: |-
    Cost Seiko API
  contact:
    email: khalil@lokalix-solutions.com
  version: 1.0.0
servers:
  - url: http://localhost:9090
...
paths:
...
  /customers/search:
    post:
      tags:
        - customer
      summary: |
        Search customers with filter:
        - freetext looking at name/code
        Return paginated data ordered alphabetically
      operationId: searchCustomers
      parameters:
        - $ref: '#/components/parameters/offsetParam'
        - $ref: '#/components/parameters/limitParam'
      requestBody:
        description: search customers
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BasicSearch'
        required: true
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomersPaginated"
  /customers/{uid}:
    parameters:
      - $ref: '#/components/parameters/uidParam'
    get:
      tags:
        - customer
      summary: Retrieve a customer
      operationId: retrieveCustomer
      responses:
        '200':
          description: customer
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'
    put:
      tags:
        - customer
      summary: Update a customer
      operationId: updateCustomer
      requestBody:
        description: update customer object
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CustomerUpdate'
        required: true
      responses:
        '200':
          description: object updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'
  /customers:
    get:
      tags:
        - customer
      summary: List all customers
      operationId: listAllCustomers
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Customer'
        '400':
          $ref: '#/components/responses/BadRequest'
    post:
      tags:
        - customer
      summary: Create a new customer
      operationId: createCustomer
      requestBody:
        description: New customer
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CustomerCreate'
        required: true
      responses:
        '201':
          description: object created
          headers:
            Uid:
              schema:
                $ref: '#/components/schemas/Uid'
              description: uid of the created object.
  /customers/{uid}/shipment-locations:
    parameters:
      - $ref: '#/components/parameters/uidParam'
    get:
      tags:
        - customer
      summary: Retrieve a customer shipment locations
      operationId: retrieveCustomerShipmentLocations
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CustomerShipmentLocation'
    post:
      tags:
        - customer
      summary: Create a customer shipment location
      operationId: createCustomerShipmentLocation
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CustomerShipmentLocationCreate'
        required: true
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CustomerShipmentLocation'
  /customers/{uid}/shipment-locations/{shipmentLocationUid}:
    parameters:
      - $ref: '#/components/parameters/uidParam'
      - $ref: '#/components/parameters/shipmentLocationUidParam'
    get:
      tags:
        - customer
      summary: Retrieve a customer shipment location
      operationId: retrieveCustomerShipmentLocation
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerShipmentLocation'
    put:
      tags:
        - customer
      summary: Update a customer shipment location
      operationId: updateCustomerShipmentLocation
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CustomerShipmentLocationCreate'
        required: true
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CustomerShipmentLocation'
  /customers/{uid}/shipment-locations/{shipmentLocationUid}/delete:
    parameters:
      - $ref: '#/components/parameters/uidParam'
      - $ref: '#/components/parameters/shipmentLocationUidParam'
    post:
      tags:
        - customer
      summary: Delete a customer shipment location
      operationId: deleteCustomerShipmentLocation
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CustomerShipmentLocation'
  /customers/{uid}/archive:
    parameters:
      - $ref: '#/components/parameters/uidParam'
    post:
      tags:
        - customer
      summary: Archive a customer
      operationId: archiveCustomer
      responses:
        '200':
          description: successful operation
  /customers/{uid}/terms-and-conditions/non-dyson:
    parameters:
      - $ref: '#/components/parameters/uidParam'
    get:
      tags:
        - terms-and-conditions
      summary: Get terms and conditions for non-dyson
      operationId: getTermsAndConditionsNonDyson
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TermsAndConditionsNonDyson'
    patch:
      tags:
        - terms-and-conditions
      summary: patch terms and conditions for non-dyson
      operationId: patchTermsAndConditionsNonDyson
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TermsAndConditionsNonDysonPatch'
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TermsAndConditionsNonDyson'
  /customers/{uid}/terms-and-conditions/dyson:
    parameters:
      - $ref: '#/components/parameters/uidParam'
    get:
      tags:
        - terms-and-conditions
      summary: Get terms and conditions for dyson
      operationId: getTermsAndConditionsDyson
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TermsAndConditionsDyson'
    patch:
      tags:
        - terms-and-conditions
      summary: patch terms and conditions for dyson
      operationId: patchTermsAndConditionsDyson
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TermsAndConditionsDysonPatch'
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TermsAndConditionsDyson'
  /customers/{uid}/requestor-names:
    parameters:
      - $ref: '#/components/parameters/uidParam'
    get:
      tags:
        - customer
      summary: retrieve requestor names from customer
      operationId: retrieveRequestorNames
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
    put:
      tags:
        - customer
      summary: set requestor names to customer
      operationId: setRequestorNames
      requestBody:
        description: requestor names
        content:
          application/json:
            schema:
              type: array
              items:
                type: string
        required: true
      responses:
        '200':
          description: successful operation
  /customers/{uid}/add-requestor-names:
    parameters:
      - $ref: '#/components/parameters/uidParam'
    post:
      tags:
        - customer
      summary: add requestor names to customer
      operationId: addRequestorNames
      requestBody:
        description: requestor names
        content:
          application/json:
            schema:
              type: array
              items:
                type: string
        required: true
      responses:
        '200':
          description: successful operation
  /customers/{uid}/customer-emails:
    parameters:
      - $ref: '#/components/parameters/uidParam'
    get:
      tags:
        - customer
      summary: retrieve customer emails from customer
      operationId: retrieveCustomerEmails
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
    put:
      tags:
        - customer
      summary: set customer emails to customer
      operationId: setCustomerEmails
      requestBody:
        description: customer emails
        content:
          application/json:
            schema:
              type: array
              items:
                type: string
        required: true
      responses:
        '200':
          description: successful operation
  /customers/{uid}/add-customer-emails:
    parameters:
      - $ref: '#/components/parameters/uidParam'
    post:
      tags:
        - customer
      summary: add customer emails to customer
      operationId: addCustomerEmails
      requestBody:
        description: customer emails
        content:
          application/json:
            schema:
              type: array
              items:
                type: string
        required: true
      responses:
        '200':
          description: successful operation
...
components:
  parameters:
    uidParam:
      in: path
      name: uid
      required: true
      schema:
        $ref: '#/components/schemas/Uid'
    offsetParam:
      in: query
      name: offset
      required: true
      schema:
        type: integer
        minimum: 0
        default: 0
      description: The number of items to skip before starting to collect the result set.
    limitParam:
      in: query
      name: limit
      required: true
      schema:
        type: integer
        minimum: 1
        maximum: 50
        default: 20
      description: The numbers of items to return.
...
    shipmentLocationUidParam:
      in: path
      name: shipmentLocationUid
      required: true
      schema:
        $ref: '#/components/schemas/Uid'
      description: The shipment location UID
...
  schemas:
...
    TermsAndConditionsNonDyson:
      properties:
        validityNumberOfDays:
          type: integer
        deliveryCharges:
          type: number
          format: double
        storageAcceptDeliveryNumberMonths:
          type: integer
        storageMinimumStorageFee:
          type: number
          format: double
        nonCancellationNumberWorkingDays:
          type: integer
        nonRescheduledNumberWeeks:
          type: integer
        claimsPackagingDamageNumberDays:
          type: number
          format: double
        forecastLeadTime:
          type: string
        latePaymentPenalty:
          type: number
          format: double
    TermsAndConditionsNonDysonPatch:
      allOf:
        - $ref: '#/components/schemas/TermsAndConditionsNonDyson'
    TermsAndConditionsDyson:
      properties:
        validityNumberOfDays:
          type: integer
        currencyExchangeRate:
          type: number
          format: double
        minimumDeliveryQuantity:
          type: integer
        storageAcceptDeliveryNumberMonths:
          type: integer
        storageMinimumStorageFee:
          type: number
          format: double
        nonCancellationNumberWorkingDays:
          type: integer
        nonRescheduledNumberWeeks:
          type: integer
        claimsPackagingDamageNumberDays:
          type: number
          format: double
        forecastLeadTime:
          type: string
        latePaymentPenalty:
          type: number
          format: double
    TermsAndConditionsDysonPatch:
      allOf:
        - $ref: '#/components/schemas/TermsAndConditionsDyson'
    PaginationInfos:
      type: object
      required:
        - total
        - page
        - per_page
        - has_next
        - has_prev
      properties:
        total: { type: integer }
        page: { type: integer }
        per_page: { type: integer }
        has_next: { type: boolean }
        has_prev: { type: boolean }
    BasicSearch:
      type: object
      required:
        - searchText
      properties:
        searchText:
          type: string
...
    Uid:
      type: string
      format: uuid
...
    CustomersPaginated:
      type: object
      allOf:
        - $ref: "#/components/schemas/PaginationInfos"
        - type: object
          required:
            - results
          properties:
            results:
              type: array
              items:
                $ref: "#/components/schemas/Customer"
    CustomerCreate:
      type: object
      required:
        - name
        - code
        - dyson
      properties:
        name:
          type: string
        code:
          type: string
        dyson:
          type: boolean
        markup:
          type: number
          format: double
        paymentTerms:
          type: string
    CustomerUpdate:
      type: object
      required:
        - name
        - code
        - dyson
      properties:
        name:
          type: string
        code:
          type: string
        dyson:
          type: boolean
        markup:
          type: number
          format: double
        paymentTerms:
          type: string
    Customer:
      type: object
      required:
        - uid
        - name
        - code
        - dyson
      properties:
        uid:
          $ref: '#/components/schemas/Uid'
        name:
          type: string
        code:
          type: string
        dyson:
          type: boolean
        markup:
          type: number
          format: double
        paymentTerms:
          type: string
        nbShipmentLocations:
          type: integer
...
    Currency:
      type: object
      required:
        - uid
        - code
        - exchangeRates
        - creationDate
      properties:
        uid:
          $ref: '#/components/schemas/Uid'
        code:
          type: string
        exchangeRates:
          type: array
          items:
            $ref: '#/components/schemas/ExchangeRate'
        creationDate:
          type: string
          format: date-time
    ExchangeRate:
      type: object
      required:
        - toCurrencyCode
        - rate
        - creationDate
      properties:
        toCurrencyCode:
          type: string
          description: "Currency code (e.g., USD, EUR, MYR)"
        rate:
          type: number
          format: double
          description: "1 unit of base currency equals <rate> units of target currency"
        creationDate:
          type: string
          format: date-time
...
    CustomerShipmentLocationCreate:
      type: object
      required:
        - shipmentLocationId
        - currencyId
      properties:
        shipmentLocationId:
          $ref: '#/components/schemas/Uid'
        currencyId:
          $ref: '#/components/schemas/Uid'
    CustomerShipmentLocation:
      type: object
      required:
        - uid
        - shipmentLocation
        - currency
      properties:
        uid:
          $ref: '#/components/schemas/Uid'
        shipmentLocation:
          $ref: '#/components/schemas/ShipmentLocation'
        currency:
          $ref: '#/components/schemas/Currency'
...
    ShipmentLocation:
      type: object
      required:
        - uid
        - name
      properties:
        uid:
          $ref: '#/components/schemas/Uid'
        name:
          type: string
        acceptedCurrencies:
          type: array
          items:
            $ref: '#/components/schemas/Currency'
...
    CustomError:
      type: object
      required:
        - code
      properties:
        code:
          $ref: '#/components/schemas/CustomErrorCode'
        message:
          type: string
    CustomErrorCode:
      type: string
      enum:
        - ACCESS_FORBIDDEN
        - INTERNAL_ERROR
        - INVALID_EMAIL_OR_PASSWORD
        - ENTITY_NOT_FOUND
        - ENTITY_ALREADY_EXISTS
        - LICENSE_LIMIT_EXCEEDED
        - USER_IS_INACTIVE
        - USER_RESET_PASSWORD_INVALID_TOKEN
        - USER_CANNOT_DEACTIVATE_ONESELF
        - FILE_UNSUPPORTED
        - FILE_EXISTS
        - FILE_NOT_FOUND
        - GENERIC_ERROR
        - MUST_BE_DISCONNECTED
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/CustomError'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT    # optional, arbitrary value for documentation purposes

# 2) Apply the security globally to all operations
security:
  - bearerAuth: [ ]         # use the same name as above